pyproject.toml
[tool.poetry]
name = "qp266-backend"
version = "0.1.0"
description = ""
authors = ["kotorkovsciy <kotorkovsciy@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
django-cors-headers = "^4.4.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
whitenoise = {extras = ["brotli"], version = "^6.7.0"}
django-health-check = "^3.18.3"
requests = "^2.32.3"
graphene-django = "^3.2.2"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pre-commit = "^3.8.0"
django-stubs = "^5.0.4"
mypy-extensions = "^1.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = '''
(
  /(
    | .venv
    | migrations
    | config/settings
  )/
)
'''

[tool.mypy]
plugins = "mypy_django_plugin.main"
